+++ config/src/main/java/com/alibaba/nacos/config/server/modules/repository/GroupCapacityRepository.java
@@ -8,7 +8,7 @@ import org.springframework.data.repository.PagingAndSortingRepository;
  * @author paderlol
  * @date: 2020/1/18 15:49
  */
-public interface GroupCapacityRepository extends PagingAndSortingRepository<GroupCapacity, Integer>,
+public interface GroupCapacityRepository extends PagingAndSortingRepository<GroupCapacity, Long>,
     QuerydslPredicateExecutor<GroupCapacity> {
 
 }
